jump to line fragment when opening local links (#648)
Utilize #Lxxx link fragment on links to local files. If file is already
opened, existing editor will be used/focused and cursor will be moved to
line specified in link fragment.
For that we use openTextDocument and showTextDocument instead of
vscode.open. Also we use Tab API that is available from VS Code v1.67.

Resolves shd101wyy/vscode-markdown-preview-enhanced#647